RDKB-30840, RDKB-30843 : Fix warnings in CcspCommonLibrary
Reason for change: To remove warnings during CcspLogAgent compilation
Removed warnings are
1. Comparison between pointer and integer [enabled by default]
2. Macro redefined
3. gethostbyname is obsolescent
Test Procedure: Build should be success, basic functionalities should be tested.
Risks: High

RDKB-30842 : Warning error in CcspCr

Reason for change: To resolve warnings in CcspCr
Test Procedure: Verify that there are no warning messages.
Risks: High


RDKB-30840: Fix warnings in CcspCommonLibrary

Reason for change: To remove the warnings
1. passsing argument discards 'const' qualifier from pointer target type [enabled by default].
2. assignment discards 'const' qualifier from pointer target type [enabled by default].
3. 'return' with a value, in function returning void [enabled by default].
4. extra tokens at end of #ifndef directive [enabled by default]
5. cast to pointer from integer of different size
6. this decimal constant is unsigned only in ISO C90 [enabled by default]
7. return makes pointer from integer without a cast [enabled by default].
8. assignment makes pointer from integer without a cast [enabled by default].
9. comparison of distinct pointer types lacks a cast [enabled by default].
10. passing argument makes integer from pointer without a cast [enabled by default].
11. comparison between pointer and integer
12. "_GNU_SOURCE" redefined .
13. passing argument from incompatible pointer type [enabled by default].
14. initialization from incompatible pointer type [enabled by default].
15. assignment from incompatible pointer type.

Test Procedure: basic functionalities should be tested, regression test should pass
Risks: High


RDKB-31127: Fix warnings in CcspCommonLibrary for AXB6.

Reason for change: To remove the warnings
1.implicit declaration of function [-Wimplicit-function-declaration]
2.conflicting types for 'UserSetDelta'
3."PAGE_SIZE" redefined
4.inline function <function name> declared but never defined
Test Procedure: Build should be successful.
Risks: High


RDKB-31156: Fix pointer/integer type mismatch warning

Reason for change: warning: pointer/integer type mismatch in conditional expression in slap_definitions.h
Test Procedure: Check if recipe builds fine without above warning.
Risks: High

(cherry picked from commit df4cf7dd2d0687a923cad248d36da974a6b634d8)
(cherry picked from commit 03af5ff268e9878b7dbc70807aac88ae82de5cda)

RDKB-30840: Fix warnings for CcspCommonLibrary

Reason for change: to fix below warnings
1. implicit declaration of function <func Name>
2. unused variable <variable name>
3. unused parameter <parameter Name>
4. variable '<variable name>' set but not used
5. missing braces around initializer
6. near initialization for 'MappedIpAddress.Dot'
7. pointer targets in passing argument 1 of 'strlen' differ in signedness
8. comparison between signed and unsigned integer expressions
9. format '%X' expects argument of type 'unsigned int *', but argument 3 has type 'ULONG *'
10. pointer targets in passing argument 1 of 'AnscBase64Decode' differ in signedness
11. assignment from incompatible pointer type
12. pointer targets in passing argument 2 of 'pExtra->GetEncodedData' differ in signedness
13. pointer targets in initialization differ in signedness
14. return makes pointer from integer without a cast
15. comparison is always true due to limited range of data type
16. statement with no effect
17. the address of 'bHasFrac' will always evaluate as 'true'
18. enumeration value not handled in switch
19. pointer/integer type mismatch in conditional expression
20. assignment makes integer from pointer without a cast

Test Procedure: Build should be successful,Build should be success, basic functionalities should be tested.
Risks: High

(cherry picked from commit 481d2f4a2bda528017500275af2fb30f772c8545)

+++ source/cosa/utilities/bsp_eng/bspeng_stack_co_process.c
@@ -213,7 +213,7 @@ BspTemplateStackGetAt
 
     pVars = (PBSP_TEMPLATE_VAR_OBJECT *)pMyObject->hVars;
 
-    if (ulIndex < 0 || ulIndex >= pMyObject->Top)
+    if ((int)ulIndex < 0 || ulIndex >= pMyObject->Top)
         return NULL;
 
     return pVars[ulIndex];
@@ -513,7 +513,7 @@ BspTemplateStackGetBufferAt
     PBSP_TEMPLATE_STACK_OBJECT      pMyObject = (PBSP_TEMPLATE_STACK_OBJECT)hThisObject;
     PBSP_TEMPLATE_VAR_OBJECT        *pVars;
 
-    if (ulIndex < 0 || ulIndex >= pMyObject->Top)
+    if ((int)ulIndex < 0 || ulIndex >= pMyObject->Top)
         return NULL;
 
     pVars   = (PBSP_TEMPLATE_VAR_OBJECT *)pMyObject->hVars;
